Raise RuntimeError if a layerXML is not readable
QgsCredentialsNull is a hack as QGIS by default uses
QgsCredentialsConsole to ask user for db access credentials, when using
GUI QgsCredentialsDialog is used. None of those solutions work in a
server enviroment as it's not possible to wait for user input and the
process will simply hang.

As QgsCredentialsNull is a Python object it doesn't work with QGIS API
as it raises "TypeError: invalid result type from
QgsCredentialsNull.request()" but it's good enough at the moment, as it
'fixes' process hang issue.